Removed some unnecessary files.
The .empty file was not necessary as rebar3 will create the priv
directory for you.  It's existence was causing me some issues when
I was packaging this as an RPM using epm.  This change removes the
.empty file and the priv directory, it also adds ebin and priv
directories to the .gitignore, and removes them when you run rebar3
clean. I bumped the version to 0.5.1.