Detach from stdin when calling endwin()
Previously, even when getch() was not currently active and ncurses operation was
suspended by calling endwin(), Erlang and cecho would still process stdin,
making it unusable for external applications such as e.g. a vim instance started
as a separate process.

This commit attempts to fix this issue by allowing the stdin descriptor to be
temporarily detached from whenever `endwin()` is called. Subsequent refresh or
getch operations re-attach to the input. As a result, it is now possible to
e.g. launch an external VIM editor from a cecho-enabled Erlang application
as follows:

launch_vim() ->
	cecho:endwin(),
	Port = open_port({spawn_executable, "/usr/bin/vim"},
						[nouse_stdio, exit_status]),
	receive
	{Port, {exit_status, _RC}} ->
		cecho:refresh()
	end.